UiB/INF142/GroupAssignment/requirements.txt
dnspython==1.16.0
numpy==1.20.2
pymongo[snappy,gssapi,srv,tls]==3.11.3
python-dotenv==0.17.0
python-snappy==0.6.0
winkerberos==0.7.0
